% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsegfit.R
\name{fitted.segfit}
\alias{fitted.segfit}
\title{Extract the fitted values from a segfit object}
\usage{
\method{fitted}{segfit}(sf, concat = TRUE)
}
\arguments{
\item{sf}{The segfit object}

\item{concat}{Whether the fitted values of different segments will be
concatenated.}
}
\value{
If concat==TRUE, the fitted values of different segments
              will be concatenated into one vector. Otherwise,
              the returned value will be a list of fitted values
              of each segment.
}
\description{
Extract the fitted values from a segfit object
}
\details{
Extract the fitted values from a segfit object
}
