% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsegfit.R
\name{summary.segfit}
\alias{summary.segfit}
\title{Summarise a segfit object}
\usage{
\method{summary}{segfit}(sf)
}
\arguments{
\item{sf}{The segfit object to summarise}
}
\description{
Summarise a segfit object
}
\details{
Summarise a segfit object
}
